Updated Mobile datatype Removed unwanted spring dependencies Moved test class to its package

+++ src/main/java/com/nokia/seatmanagement/core/notification/SMSNotifier.java
@@ -15,7 +15,7 @@ public class SMSNotifier implements Notifier {
 	public void notify(Integer userId, String message) throws InvalidUserException {
 		
 		Users user = SeatManagementDbHelper.getUserDetails(userId);
-		Integer mobileNumber = user.getMobile();
+		Long mobileNumber = user.getMobile();
 		System.out.println("Mock SMS Notification sent to:" + mobileNumber + " message:"  + message);
 		
 		/*
